change types from 8 bit to 16 bit

+++ MIKO-16/MIKO_Processor.h
@@ -6,10 +6,10 @@
 class MIKO_Processor
 {
 private:
-	Uint8 reg_a = 0, reg_x = 0, reg_y = 0, reg_sp = 0xff, reg_status = 0;
+	Uint16 reg_a = 0, reg_x = 0, reg_y = 0, reg_sp = 0xff, reg_status = 0;
 	// Initialise the CPU registers
 
-	long long int reg_pc = 0;
+	long long unsigned reg_pc = 0;
 	// Initialise the CPU program counter (64 bit) a.k.a Instruction Pointer
 
 public:
